fix: signup form and useruuid variable
