+++ src/pages/api/auth/signout.ts
@@ -3,6 +3,6 @@ import { deleteCookie } from 'cookies-next';
 export default function handler() {
   // delete the token cookie
   deleteCookie('token');
-  // delete the userid from the local storage
-  localStorage.removeItem('userid');
+  // delete the userUuid from the local storage
+  localStorage.removeItem('userUuid');
 }
